backend/requirements.txt

flask
flask-cors
langchain
langchain-core
langchain-community
langchain-groq
langchain-tavily
biopython
arxiv
aiohttp
PyPDF2
chromadb
pydantic
sentence-transformers
